Creating verifyTokenAndRateLimit and modifying all classes
