--- BackEnd/src/middleware/rateLimiters.ts
@@ -1,8 +1,0 @@
-import rateLimit from 'express-rate-limit';
-
-export const createDietLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100,
-  standardHeaders: true,
-  legacyHeaders: false,
-});
