--- BackEnd/src/middleware/verifyToken.ts
@@ -1,23 +1,0 @@
-import { Request, Response, NextFunction } from 'express';
-import jwt from 'jsonwebtoken';
-
-export const verifyToken = (req: Request, res: Response, next: NextFunction) => {
-  
-  const authHeader = req.headers.authorization;
-  if (!authHeader) {
-    return res.status(401).json({ error: 'Token not provided' });
-  }
-
-  const token = authHeader.split(' ')[1];
-  try {
-    const decoded = jwt.verify(token, process.env.JWT_SECRET!);
-    if (typeof decoded === 'object' && 'id' in decoded) {
-      req.user = { id: decoded.id };  
-      next();
-    } else {
-      return res.status(401).json({ error: 'Invalid token' });
-    }
-  } catch (error) {
-    return res.status(401).json({ error: 'Invalid token' });
-  }
-};
